Proto-login file
Adding a login system with simulated backend to the prototype folder. I originally intended to merge it with the current prototype to allow both to be ran using the same packages, but different package versions made that seem impossible. To open the login page, go to the proto\proto-login directory and run 'npm start'.