View Page + Vuetify Implementation
Added an additional page after 'Home' page (or the userlist page). This new page, intended to become the student view page, only has a logout and placeholder page link button. The project was also modified to support vuetify. The "Continue" button on the home page and the placeholder button on the student view page both implement vuetify.