Popup Class Summaries
Added popup text for when a class is selected on /studentview.  This popup should display as much information about the class as possible. Two options appear: 'OK' and 'REMOVE', but they don't yet have functionality.

Also added a class list to /studentaddclass. It functions the same as the studentview list, but the two options are 'ADD CLASS' and 'CANCEL'.